Pull CustomTabsCAllback to a getter from private field
TWA based apps might want to listen to CCT call back events and take appropriate action ( eg. capture metrics about usage). Since the field currently is private and used in launchTWA() its not possible for the extending classes to change the behavior. This change pulls out the private field and creates a overridable method that supplies the Callback