Use FileData from incoming intent if it's present
Incoming intent can contain FileData extra field in the case if
LauncherActivity is relaunched in a separate task. It's related to
navigate-new launch_handler client_mode becasue it requires launch in
a new task. In this case we need to get FileData not from the data of
the intent but from the extra field.